refactor and update tests
